+ xoroshiro.h splitmix.h pcg.h wy.h photon.h romuduo.h

+++ source/combine.c
@@ -1,4 +1,4 @@
 #include "combine.h"
 uint64_t random_combine(uint32_t low, uint32_t high) {
-  return low + (high << 32);
+  return low + ((uint64_t)high << 32);
 }
